권현우, 이수빈 업데이트
If you click or drag the UI, an object will be created at the desired location. Additionally, clicking the button created in the Hierarchy window selects the current object, and clicking the same button again deselects it. If you select a different object while another object is already selected, the newly selected object will be chosen. Double-clicking an object selects it, and the object moves as the mouse moves. When you double-click, the final mouse position becomes the new location of the object.